+++ cranelift/assembler-x64/fuzz/Cargo.toml
@@ -9,8 +9,10 @@ rust-version.workspace = true
 cargo-fuzz = true
 
 [dependencies]
-libfuzzer-sys = "0.4"
-cranelift-assembler-x64 = { path = ".." }
+libfuzzer-sys = { workspace = true }
+cranelift-assembler-x64 = { path = "..", features = ['arbitrary'] }
+capstone = { workspace = true }
+arbitrary = { workspace = true, features = ['derive'] }
 
 [[bin]]
 name = "roundtrip"
@@ -18,5 +20,3 @@ path = "fuzz_targets/roundtrip.rs"
 test = false
 doc = false
 bench = false
-
-[workspace]
